Adding a key pair and security group for build servers and some other enhancements
We added a key pair specifically to connect to the build servers and created a security group to allow access to build servers from the jenkins master only. We also added lifecycle element in security group to handle renaming of security group gracefully and renamed the initial security group to jenkins_master_security group which is more descriptive. We also updated references to it.